structured the tests
